Dynamic wrapping of R functions as C callbacks

Function to wrap R functions as C function pointers.

Usage

ccallback(signature, fun, envir = new.env())

Arguments

signature

character string specifying the call signature of the C function callback type.

fun

R function to be wrapped as a C function pointer.

envir

the environment in which to evaluate the call to fun.

Details

Callbacks are user-defined functions that are registered in a foreign library and that are executed at a later time from within that library. Examples include user-interface event handlers that are registered in GUI toolkits, and, comparison functions for custom data types to be passed to generic sort algorithm.

The function ccallback wraps an R function fun as a C function pointer and returns an external pointer. The foreign C function type of the wrapped R function is specified by a call signature given by signature.

When the C function pointer is called, a global callback handler (implemented in C) is executed first, that dynamically creates an R call expression to fun using the arguments, passed from C and converted to R, according to the argument types signature within the call signature specified. See dyncall for details on the format.

Finally, the handler evaluates the R call expression within the environment given by envir. On return, the R return value of fun is coerced to the C value, according to the return type signature specified in signature. If an error occurs during the evaluation, the callback will be disabled for further invocations. (This behaviour might change in the future.)

Value

ccallback returns an external pointer to a synthetically generated C function.

Portability

The implementation is based on the dyncallback library (part of the DynCall project).

The following processor architectures are supported: X86, X64, ARM (including Thumb) and partial stable support for PowerPC 32-bit; The library has been built and tested to work on various OSs: Linux, Mac OS X, Windows 32/64-bit, BSDs, Haiku, Nexenta/Open Solaris, Minix and Plan9, as well as embedded platforms such as Linux/ARM (OpenMoko, Beagleboard, Gumstix, Efika MX, Raspberry Pi), Nintendo DS (ARM), Sony Playstation Portable (MIPS 32-bit/eabi) and iOS (ARM - armv6 mode ok, armv7 unstable). Special notes for PowerPC 32-Bit: Callbacks for System V (Linux/BSD) are unstable in this release; MacOS X/Darwin works fine. In the context of R, dyncallback has currently no support for callbacks on MIPS, SPARC and PowerPC 64-Bit. Using dyncallback to implement non-default calling conventions is not supported yet. (e.g. Window Procedures on Win32/X86).

Note

The call signature MUST match the foreign C callback function type, otherwise an activated callback call from C can lead to a fatal R process crash.

A small amount of memory is allocated with each wrapper. A finalizer function that frees the allocated memory is registered at the external pointer. If the external callback function pointer is registered in a C library, a reference should also be held in R as long as the callback can be activated from a foreign C run-time context, otherwise the garbage collector might call the finalizer and the next invocation of the callback could lead to a fatal R process crash as well.

References

Adler, D. (2012) “Foreign Library Interface”, The R Journal, 4(1), 30--40, June 2012. https://journal.r-project.org/articles/RJ-2012-004/

Adler, D., Philipp, T. (2008) DynCall Project. https://dyncall.org

See also

Examples

# \donttest{
# Create a function, wrap it to a callback and call it via dyncall:
f <- function(x, y) x + y
cb <- ccallback("ii)i", f)
r <- dyncall(cb, "ii)i", 20, 3)

# Sort vectors directly via 'qsort' C library function using an R callback:
dynbind(c("msvcrt","c","c.so.6"), "qsort(piip)v;")
#> $libhandle
#> <pointer: 0x7f194b29cc30>
#> attr(,"path")
#> [1] "libc.so.6"
#> attr(,"auto.unload")
#> [1] TRUE
#> 
#> $unresolved.symbols
#> character(0)
#> 
#> attr(,"class")
#> [1] "dynbind.report"
cb <- ccallback("pp)i", function(px, py) {
    x <- unpack(px, 0, "d")
    y <- unpack(py, 0, "d")
    if (x >  y) return(1) else if (x == y) return(0) else return(-1)
})
x <- rnorm(100)
qsort(x, length(x), 8, cb)
#> NULL
x
#>   [1] -2.38522244 -1.86449352 -1.73451441 -1.55967487 -1.54382468 -1.53294451
#>   [7] -1.34708908 -1.24717386 -1.22305337 -1.21405299 -1.12616199 -1.12123984
#>  [13] -1.02031104 -1.01033483 -0.91128224 -0.90487151 -0.88192189 -0.87465293
#>  [19] -0.86490339 -0.84939789 -0.82037232 -0.80556315 -0.78483523 -0.68892789
#>  [25] -0.68106088 -0.67930166 -0.65168158 -0.62833235 -0.61350944 -0.46258421
#>  [31] -0.44257700 -0.40316910 -0.34370481 -0.31070892 -0.27736100 -0.27719334
#>  [37] -0.21964153 -0.20894433 -0.20204341 -0.18708519 -0.18157073 -0.11765854
#>  [43] -0.10011418 -0.08970302 -0.08376107 -0.07493962 -0.01768816 -0.01335815
#>  [49]  0.02250571  0.02587341  0.18924302  0.21431505  0.24595023  0.29974584
#>  [55]  0.36940352  0.37029475  0.37133409  0.39593934  0.39946943  0.45126832
#>  [61]  0.49494114  0.50556534  0.54800558  0.56433378  0.57087680  0.57515860
#>  [67]  0.62231350  0.62634964  0.65026012  0.67143198  0.71612085  0.75210885
#>  [73]  0.75915697  0.77348994  0.92669094  0.93283383  0.94238656  0.94240723
#>  [79]  0.95690820  0.96934524  0.97011204  1.09225667  1.10569415  1.11940719
#>  [85]  1.12437391  1.22427347  1.27989404  1.28190400  1.37875474  1.41423968
#>  [91]  1.48649589  1.49437156  1.50203471  1.61782226  1.74695088  1.92206047
#>  [97]  2.08370032  2.14427855  2.26984807  2.44312061
# }