rdyncall demos: Platform installation notes for required libraries

The demos of the rdyncall package (see demo(package="rdyncall")) use shared libraries such as SDL, OpenGL and Expat via dynports - a dynamic binding approach which requires, that prebuilt binary shared library files are already installed.

Depending on the host system, some libraries are officially a part of the OS or Distribution, some others need to be installed to get the demos running.

As of the current version of this package, the installation of additional shared C libraries need to be done manually. It follows an overview of the required libraries and installation notes for various operating-systems and distributions.

Overview of Libraries

The following Libraries are used as 'run-time' pre-compiled binaries for the particular target OS and Hardware platform. Some notes on installation of additional run-time libraries required for some rdyncall demos:

Lib	Description	URL

expat	XML Parser	http://www.libexpat.org

GL	Open Graphics Library	http://opengl.org, http://www.mesa3d.org

GLU	OpenGL Utility Library	see links above

glew	OpenGL Extension Wrangler Library	https://glew.sourceforge.net/

SDL	Multimedia Framework	http://libsdl.org/

SDL_mixer	Music Format playing	http://www.libsdl.org/projects/SDL_mixer/

SDL_image	Image Format loading	http://www.libsdl.org/projects/SDL_image/

SDL_ttf	True Type Font rendering	http://www.libsdl.org/projects/SDL_ttf/

SDL_net	Network I/O	http://www.libsdl.org/projects/SDL_net/

In short: Place the shared libraries ( *.DLL, *.so or *.dylib) in a standard location or modify LD_LIBRARY_PATH(unix) or PATH(windows) so that dynfind can find the libraries.

On Mac OS X framework folders are supported as well. Place the *.framework folder at /Library/Frameworks.

Detailed platform-specific installation instructions follow up.

Windows Installation Notes

Download the *.zip files, unpack them and place the *.DLL files to a place within PATH.

32-Bit versions:

Lib	Download Link

expat	https://expat.sourceforge.net (TODO:test installer)

GL	pre-installed

GLU	pre-installed

glew	http://sourceforge.net/projects/glew/files/glew/1.7.0/glew-1.7.0-win32.zip/download

SDL	http://www.libsdl.org/release/SDL-1.2.14-win32.zip

SDL_image	http://www.libsdl.org/projects/SDL_image/release/SDL_image-1.2.10-win32.zip

SDL_mixer	http://www.libsdl.org/projects/SDL_mixer/release/SDL_mixer-1.2.11-win32.zip

SDL_ttf	http://www.libsdl.org/projects/SDL_ttf/release/SDL_ttf-2.0.10-win32.zip

SDL_net	http://www.libsdl.org/projects/SDL_net/release/SDL_net-1.2.7-win32.zip

64-Bit version:

Lib	Downdload Link

expat	no prebuilt found (TODO: build)

GL	pre-installed

GLU	pre-installed

glew	http://sourceforge.net/projects/glew/files/glew/1.7.0/glew-1.7.0-win64.zip/download

SDL	http://mamedev.org/tools/20100102/sdl-1.2.14-r5428-w64.zip

SDL_image	pre-built n/a

SDL_mixer	pre-built n/a

SDL_ttf	pre-built n/a

SDL_net	pre-built n/a

The prebuilt version of SDL from http://www.drangon.org/mingw did not work (exiting with OpenGL errors). If you know of other resources for prebuilt 64-bit packages for SDL and expat, please report.

Mac OS X Installation Notes

Download the *.dmg files, mount them (by double-click) and copy *.framework folders to /Library/Frameworks.

Lib	Download link

expat	pre-installed

GL	pre-installed

GLU	pre-installed

glew	port install glew

SDL	http://www.libsdl.org/release/SDL-1.2.14.dmg

SDL_image	http://www.libsdl.org/projects/SDL_image/release/SDL_image-1.2.10.dmg

SDL_mixer	http://www.libsdl.org/projects/SDL_mixer/release/SDL_mixer-1.2.11.dmg

SDL_ttf	http://www.libsdl.org/projects/SDL_ttf/release/SDL_ttf-2.0.10.dmg

SDL_net	http://www.libsdl.org/projects/SDL_net/release/SDL_net-1.2.7.dmg

Linux/Debian Installation Notes

Debian Package installation via aptitude

aptitude install <pkg-names>..

Lib	Debian Package name(s)

expat	libexpat1 (version 1.5.2 - already installed?)

GL	libgl1-mesa-glx and libgl1-mesa-dri

GLU	libglu1-mesa

glew	libglew1.5

SDL	libsdl1.2debian and libsdl1.2debian-<SOUNDSYS>

SDL_image	libsdl-image1.2

SDL_mixer	libsdl-mixer1.2

SDL_ttf	libsdl-ttf2.0

SDL_net	libsdl-net1.2

Depending on your sound system, <SOUNDSYS> should be explaced with one of the following: alsa, all, esd, arts, oss, nas or pulseaudio. Tested with Debian 5 and 6 (lenny and squeeze).

Linux/Fedora Installation Notes

pkcon install <pkgname>..

Lib	RPM Package name

expat	expat

GL	mesa-libGL

GLU	mesa-libGLU

glew	glew

SDL	SDL

SDL_image	SDL_image

SDL_mixer	SDL_mixer

SDL_ttf	SDL_ttf

SDL_net	SDL_net

Tested with Fedora 13 and 14 on x86 and x86_64.

Linux/openSUSE Installation Notes

zypper in <pkgname>..

Lib	Package Name

SDL	libSDL

SDL_image	libSDL_image

SDL_mixer	libSDL_mixer

SDL_net	libSDL_net

SDL_ttf	libSDL_ttf

glew	libGLEW1_6

openSUSE installation notes have not been confirmed.

NetBSD Installation Notes

Installation via pkgsrc:

pkg_add <pkgname>..

Lib	pkgsrc name

expat	expat

GL	Mesa

GLU	glu

glew	glew

SDL	SDL

SDL_image	SDL_image

SDL_mixer	SDL_mixer

SDL_ttf	SDL_ttf

SDL_net	SDL_net

OpenBSD Installation Notes

Using packages:

pkg_add <pkgname>..

Lib	port name

expat	expat

SDL	SDL

SDL_image	sdl-image

SDL_mixer	sdl-mixer

SDL_ttf	not available

SDL_net	sdl-net

The SDL dynport failed on OpenBSD 4.8 - so no multimedia demos here - using the R 2.7 from the ports tree. This must have been something to do with pthread discrepancies between SDL and R.

FreeBSD Installation Notes

Using packages:

pkg_add -r <pkgname>..

Lib	pkgname

expat	expat2

GL	xorg

glew	glew

SDL	sdl

SDL_image	sdl_image

SDL_mixer	sdl_mixer

SDL_ttf	sdl_ttf

SDL_net	sdl_net

Solaris Installation Notes

OpenCSW offers prebuilt binaries for Solaris. The installation of OpenCSW packages is done via pkgutil.

pkgutil -i <pkgname>..

See http://www.opencsw.org for details on the OpenCSW project.

Lib	pkgname

expat	expat

GL	mesalibs

GLU	mesalibs

glew	glew

SDL	libsdl1_2_0

SDL_image	sdlimage

SDL_mixer	sdlmixer

SDL_net	sdlnet

SDL_ttf	sdlttf